+++ test/io_tests/request_response_tests/authenticate_test.py
@@ -2,6 +2,9 @@
 
 from unittest.mock import Mock, patch
 
+import pytest
+
+from xknx.exceptions import RequestResponseError
 from xknx.io.request_response import Authenticate
 from xknx.knxip import (
     HPAI,
@@ -28,7 +31,7 @@ class TestAuthenticate:
         )
         authenticate.timeout_in_seconds = 0
 
-        assert authenticate.awaited_response_class == SessionStatus
+        assert authenticate.AWAITED_RESPONSE_CLASS is SessionStatus
 
         # Expected KNX/IP-Frame:
         exp_knxipframe = KNXIPFrame.init_from_body(
@@ -38,24 +41,29 @@ class TestAuthenticate:
             )
         )
 
-        await authenticate.start()
+        with pytest.raises(RequestResponseError):
+            await authenticate.request()
         transport_mock.send.assert_called_with(exp_knxipframe)
 
         # Response KNX/IP-Frame with wrong type
         wrong_knxipframe = KNXIPFrame.init_from_body(ConnectionStateRequest())
         with patch("logging.Logger.warning") as mock_warning:
-            authenticate.response_rec_callback(wrong_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with("Could not understand knxipframe")
-            assert authenticate.success is False
+            authenticate._response_rec_callback(wrong_knxipframe, HPAI(), None)
+            mock_warning.assert_called_with(
+                "Could not understand knxipframe for %s: %s",
+                type(authenticate).__name__,
+                wrong_knxipframe,
+            )
+            assert authenticate._response is None
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(SessionStatus())
         res_knxipframe.body.status = (
             SecureSessionStatusCode.STATUS_AUTHENTICATION_SUCCESS
         )
-        authenticate.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert authenticate.success is True
+        authenticate._response_rec_callback(res_knxipframe, HPAI(), None)
+        assert authenticate._response is not None
         assert (
-            authenticate.response.status
+            authenticate._response.status
             == SecureSessionStatusCode.STATUS_AUTHENTICATION_SUCCESS
         )