+++ test/io_tests/request_response_tests/connect_test.py
@@ -2,6 +2,9 @@
 
 from unittest.mock import patch
 
+import pytest
+
+from xknx.exceptions import RequestResponseError
 from xknx.io.request_response import Connect
 from xknx.io.transport import UDPTransport
 from xknx.knxip import (
@@ -26,7 +29,7 @@ class TestConnect:
         connect = Connect(udp_transport, local_hpai=local_hpai)
         connect.timeout_in_seconds = 0
 
-        assert connect.awaited_response_class == ConnectResponse
+        assert connect.AWAITED_RESPONSE_CLASS is ConnectResponse
 
         # Expected KNX/IP-Frame:
         exp_knxipframe = KNXIPFrame.init_from_body(
@@ -41,27 +44,27 @@ class TestConnect:
             patch("xknx.io.transport.UDPTransport.getsockname") as mock_udp_getsockname,
         ):
             mock_udp_getsockname.return_value = ("192.168.1.3", 4321)
-            await connect.start()
+            with pytest.raises(RequestResponseError):
+                await connect.request()
             mock_udp_send.assert_called_with(exp_knxipframe)
 
         # Response KNX/IP-Frame with wrong type
         wrong_knxipframe = KNXIPFrame.init_from_body(ConnectionStateRequest())
         with patch("logging.Logger.warning") as mock_warning:
-            connect.response_rec_callback(wrong_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with("Could not understand knxipframe")
+            connect._response_rec_callback(wrong_knxipframe, HPAI(), None)
+            mock_warning.assert_called_with(
+                "Could not understand knxipframe for %s: %s",
+                type(connect).__name__,
+                wrong_knxipframe,
+            )
 
         # Response KNX/IP-Frame with error:
         err_knxipframe = KNXIPFrame.init_from_body(
             ConnectResponse(status_code=ErrorCode.E_CONNECTION_ID)
         )
-        with patch("logging.Logger.debug") as mock_warning:
-            connect.response_rec_callback(err_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with(
-                "Error: KNX bus responded to request of type '%s' with error in '%s': %s",
-                type(connect).__name__,
-                type(err_knxipframe.body).__name__,
-                ErrorCode.E_CONNECTION_ID,
-            )
+        connect._response_rec_callback(err_knxipframe, HPAI(), None)
+        assert connect._response is None
+        assert connect._error_code is ErrorCode.E_CONNECTION_ID
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(
@@ -70,10 +73,10 @@ class TestConnect:
                 crd=ConnectResponseData(individual_address=IndividualAddress(7)),
             )
         )
-        connect.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert connect.success
-        assert connect.communication_channel == 23
-        assert connect.crd.individual_address.raw == 7
+        connect._response_rec_callback(res_knxipframe, HPAI(), None)
+        assert connect._response is not None
+        assert connect._response.communication_channel == 23
+        assert connect._response.crd.individual_address.raw == 7
 
     async def test_connect_route_back_true(self) -> None:
         """Test connecting from KNX bus."""
@@ -82,7 +85,7 @@ class TestConnect:
         connect = Connect(udp_transport, local_hpai=local_hpai)
         connect.timeout_in_seconds = 0
 
-        assert connect.awaited_response_class == ConnectResponse
+        assert connect.AWAITED_RESPONSE_CLASS is ConnectResponse
 
         # Expected KNX/IP-Frame:
         exp_knxipframe = KNXIPFrame.init_from_body(ConnectRequest())
@@ -91,27 +94,27 @@ class TestConnect:
             patch("xknx.io.transport.UDPTransport.getsockname") as mock_udp_getsockname,
         ):
             mock_udp_getsockname.return_value = ("192.168.1.3", 4321)
-            await connect.start()
+            with pytest.raises(RequestResponseError):
+                await connect.request()
             mock_udp_send.assert_called_with(exp_knxipframe)
 
         # Response KNX/IP-Frame with wrong type
         wrong_knxipframe = KNXIPFrame.init_from_body(ConnectionStateRequest())
         with patch("logging.Logger.warning") as mock_warning:
-            connect.response_rec_callback(wrong_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with("Could not understand knxipframe")
+            connect._response_rec_callback(wrong_knxipframe, HPAI(), None)
+            mock_warning.assert_called_with(
+                "Could not understand knxipframe for %s: %s",
+                type(connect).__name__,
+                wrong_knxipframe,
+            )
 
         # Response KNX/IP-Frame with error:
         err_knxipframe = KNXIPFrame.init_from_body(
             ConnectResponse(status_code=ErrorCode.E_CONNECTION_ID)
         )
-        with patch("logging.Logger.debug") as mock_warning:
-            connect.response_rec_callback(err_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with(
-                "Error: KNX bus responded to request of type '%s' with error in '%s': %s",
-                type(connect).__name__,
-                type(err_knxipframe.body).__name__,
-                ErrorCode.E_CONNECTION_ID,
-            )
+        connect._response_rec_callback(err_knxipframe, HPAI(), None)
+        assert connect._response is None
+        assert connect._error_code is ErrorCode.E_CONNECTION_ID
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(
@@ -120,7 +123,7 @@ class TestConnect:
                 crd=ConnectResponseData(individual_address=IndividualAddress(7)),
             )
         )
-        connect.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert connect.success
-        assert connect.communication_channel == 23
-        assert connect.crd.individual_address.raw == 7
+        connect._response_rec_callback(res_knxipframe, HPAI(), None)
+        assert connect._response is not None
+        assert connect._response.communication_channel == 23
+        assert connect._response.crd.individual_address.raw == 7