+++ test/io_tests/request_response_tests/connectionstate_test.py
@@ -2,6 +2,9 @@
 
 from unittest.mock import patch
 
+import pytest
+
+from xknx.exceptions import RequestResponseError
 from xknx.io.request_response import ConnectionState
 from xknx.io.transport import UDPTransport
 from xknx.knxip import (
@@ -26,7 +29,7 @@ class TestConnectionState:
         )
         connectionstate.timeout_in_seconds = 0
 
-        assert connectionstate.awaited_response_class == ConnectionStateResponse
+        assert connectionstate.AWAITED_RESPONSE_CLASS is ConnectionStateResponse
         assert connectionstate.communication_channel_id == communication_channel_id
 
         # Expected KNX/IP-Frame:
@@ -41,32 +44,32 @@ class TestConnectionState:
             patch("xknx.io.transport.UDPTransport.getsockname") as mock_udp_getsockname,
         ):
             mock_udp_getsockname.return_value = ("192.168.1.3", 4321)
-            await connectionstate.start()
+            with pytest.raises(RequestResponseError):
+                await connectionstate.request()
             mock_udp_send.assert_called_with(exp_knxipframe)
 
         # Response KNX/IP-Frame with wrong type
         wrong_knxipframe = KNXIPFrame.init_from_body(ConnectionStateRequest())
         with patch("logging.Logger.warning") as mock_warning:
-            connectionstate.response_rec_callback(wrong_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with("Could not understand knxipframe")
+            connectionstate._response_rec_callback(wrong_knxipframe, HPAI(), None)
+            mock_warning.assert_called_with(
+                "Could not understand knxipframe for %s: %s",
+                type(connectionstate).__name__,
+                wrong_knxipframe,
+            )
 
         # Response KNX/IP-Frame with error:
         err_knxipframe = KNXIPFrame.init_from_body(
             ConnectionStateResponse(status_code=ErrorCode.E_CONNECTION_ID)
         )
-        with patch("logging.Logger.debug") as mock_warning:
-            connectionstate.response_rec_callback(err_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with(
-                "Error: KNX bus responded to request of type '%s' with error in '%s': %s",
-                type(connectionstate).__name__,
-                type(err_knxipframe.body).__name__,
-                ErrorCode.E_CONNECTION_ID,
-            )
+        connectionstate._response_rec_callback(err_knxipframe, HPAI(), None)
+        assert connectionstate._response is None
+        assert connectionstate._error_code is ErrorCode.E_CONNECTION_ID
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(ConnectionStateResponse())
-        connectionstate.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert connectionstate.success
+        connectionstate._response_rec_callback(res_knxipframe, HPAI(), None)
+        assert connectionstate._response is not None
 
     async def test_connectionstate_route_back_true(self) -> None:
         """Test connectionstateing from KNX bus."""
@@ -78,7 +81,7 @@ class TestConnectionState:
         )
         connectionstate.timeout_in_seconds = 0
 
-        assert connectionstate.awaited_response_class == ConnectionStateResponse
+        assert connectionstate.AWAITED_RESPONSE_CLASS is ConnectionStateResponse
         assert connectionstate.communication_channel_id == communication_channel_id
 
         # Expected KNX/IP-Frame:
@@ -92,29 +95,29 @@ class TestConnectionState:
             patch("xknx.io.transport.UDPTransport.getsockname") as mock_udp_getsockname,
         ):
             mock_udp_getsockname.return_value = ("192.168.1.3", 4321)
-            await connectionstate.start()
+            with pytest.raises(RequestResponseError):
+                await connectionstate.request()
             mock_udp_send.assert_called_with(exp_knxipframe)
 
         # Response KNX/IP-Frame with wrong type
         wrong_knxipframe = KNXIPFrame.init_from_body(ConnectionStateRequest())
         with patch("logging.Logger.warning") as mock_warning:
-            connectionstate.response_rec_callback(wrong_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with("Could not understand knxipframe")
+            connectionstate._response_rec_callback(wrong_knxipframe, HPAI(), None)
+            mock_warning.assert_called_with(
+                "Could not understand knxipframe for %s: %s",
+                type(connectionstate).__name__,
+                wrong_knxipframe,
+            )
 
         # Response KNX/IP-Frame with error:
         err_knxipframe = KNXIPFrame.init_from_body(
             ConnectionStateResponse(status_code=ErrorCode.E_CONNECTION_ID)
         )
-        with patch("logging.Logger.debug") as mock_warning:
-            connectionstate.response_rec_callback(err_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with(
-                "Error: KNX bus responded to request of type '%s' with error in '%s': %s",
-                type(connectionstate).__name__,
-                type(err_knxipframe.body).__name__,
-                ErrorCode.E_CONNECTION_ID,
-            )
+        connectionstate._response_rec_callback(err_knxipframe, HPAI(), None)
+        assert connectionstate._response is None
+        assert connectionstate._error_code is ErrorCode.E_CONNECTION_ID
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(ConnectionStateResponse())
-        connectionstate.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert connectionstate.success
+        connectionstate._response_rec_callback(res_knxipframe, HPAI(), None)
+        assert connectionstate._response is not None