+++ test/io_tests/request_response_tests/tunnelling_test.py
@@ -2,8 +2,11 @@
 
 from unittest.mock import patch
 
+import pytest
+
 from xknx.cemi import CEMIFrame, CEMILData, CEMIMessageCode
 from xknx.dpt import DPTArray
+from xknx.exceptions import RequestResponseError
 from xknx.io.request_response import Tunnelling
 from xknx.io.transport import UDPTransport
 from xknx.knxip import (
@@ -42,7 +45,7 @@ class TestTunnelling:
         tunnelling = Tunnelling(udp_transport, data_endpoint, tunnelling_request)
         tunnelling.timeout_in_seconds = 0
 
-        assert tunnelling.awaited_response_class == TunnellingAck
+        assert tunnelling.AWAITED_RESPONSE_CLASS is TunnellingAck
 
         exp_knxipframe = KNXIPFrame.init_from_body(tunnelling_request)
         with (
@@ -50,29 +53,29 @@ class TestTunnelling:
             patch("xknx.io.transport.UDPTransport.getsockname") as mock_udp_getsockname,
         ):
             mock_udp_getsockname.return_value = ("192.168.1.3", 4321)
-            await tunnelling.start()
+            with pytest.raises(RequestResponseError):
+                await tunnelling.request()
             mock_udp_send.assert_called_with(exp_knxipframe, addr=data_endpoint)
 
         # Response KNX/IP-Frame with wrong type
         wrong_knxipframe = KNXIPFrame.init_from_body(ConnectionStateRequest())
         with patch("logging.Logger.warning") as mock_warning:
-            tunnelling.response_rec_callback(wrong_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with("Could not understand knxipframe")
+            tunnelling._response_rec_callback(wrong_knxipframe, HPAI(), None)
+            mock_warning.assert_called_with(
+                "Could not understand knxipframe for %s: %s",
+                type(tunnelling).__name__,
+                wrong_knxipframe,
+            )
 
         # Response KNX/IP-Frame with error:
         err_knxipframe = KNXIPFrame.init_from_body(
             TunnellingAck(status_code=ErrorCode.E_CONNECTION_ID)
         )
-        with patch("logging.Logger.debug") as mock_warning:
-            tunnelling.response_rec_callback(err_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with(
-                "Error: KNX bus responded to request of type '%s' with error in '%s': %s",
-                type(tunnelling).__name__,
-                type(err_knxipframe.body).__name__,
-                ErrorCode.E_CONNECTION_ID,
-            )
+        tunnelling._response_rec_callback(err_knxipframe, HPAI(), None)
+        assert tunnelling._response is None
+        assert tunnelling._error_code is ErrorCode.E_CONNECTION_ID
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(TunnellingAck())
-        tunnelling.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert tunnelling.success
+        tunnelling._response_rec_callback(res_knxipframe, HPAI(), None)
+        assert tunnelling._response is not None