+++ xknx/io/request_response/connect.py
@@ -9,7 +9,6 @@ from xknx.knxip import (
     ConnectRequest,
     ConnectRequestInformation,
     ConnectResponse,
-    ConnectResponseData,
     KNXIPFrame,
 )
 
@@ -19,12 +18,14 @@ if TYPE_CHECKING:
     from xknx.io.transport import KNXIPTransport
 
 
-class Connect(RequestResponse):
+class Connect(RequestResponse[ConnectResponse]):
     """
     Class to send a ConnectRequest and wait for ConnectResponse.
 
     Setting a `individual_address` is only supported for Tunnelling v2 connections.
     """
+
+    __slots__ = ("cri", "local_hpai")
 
     def __init__(
         self,
@@ -33,14 +34,11 @@ class Connect(RequestResponse):
         cri: ConnectRequestInformation | None = None,
     ) -> None:
         """Initialize Connect class."""
-        super().__init__(transport, ConnectResponse)
-        self.communication_channel = 0
-        self.data_endpoint = HPAI()
+        super().__init__(transport)
         self.local_hpai = local_hpai
         self.cri = cri or ConnectRequestInformation()
-        self.crd = ConnectResponseData()
 
-    def create_knxipframe(self) -> KNXIPFrame:
+    def _create_knxipframe(self) -> KNXIPFrame:
         """Create KNX/IP Frame object to be sent to device."""
         # use the same HPAI for control_endpoint and data_endpoint
         connect_request = ConnectRequest(
@@ -49,10 +47,3 @@ class Connect(RequestResponse):
             cri=self.cri,
         )
         return KNXIPFrame.init_from_body(connect_request)
-
-    def on_success_hook(self, knxipframe: KNXIPFrame) -> None:
-        """Set communication channel and identifier after having received a valid answer."""
-        assert isinstance(knxipframe.body, ConnectResponse)
-        self.communication_channel = knxipframe.body.communication_channel
-        self.data_endpoint = knxipframe.body.data_endpoint
-        self.crd = knxipframe.body.crd