+++ xknx/io/request_response/device_configuration.py
@@ -13,10 +13,12 @@ if TYPE_CHECKING:
     from xknx.io.transport import UDPTransport
 
 
-class DeviceConfiguration(RequestResponse):
+class DeviceConfiguration(RequestResponse[DeviceConfigurationAck]):
     """Class to send DeviceConfigurationRequest and wait for DeviceConfigurationAck (UDP only)."""
 
-    transport: UDPTransport
+    __slots__ = ("data_endpoint_addr", "device_configuration_request")
+
+    _transport: UDPTransport
 
     def __init__(
         self,
@@ -28,14 +30,12 @@ class DeviceConfiguration(RequestResponse):
         """Initialize DeviceConfiguration class."""
         self.data_endpoint_addr = data_endpoint
         self.device_configuration_request = device_configuration_request
-        super().__init__(
-            transport, DeviceConfigurationAck, timeout_in_seconds=timeout_in_seconds
-        )
+        super().__init__(transport, timeout_in_seconds=timeout_in_seconds)
 
-    async def send_request(self) -> None:
+    async def _send_request(self) -> None:
         """Build knxipframe (within derived class) and send via UDP."""
-        self.transport.send(self.create_knxipframe(), addr=self.data_endpoint_addr)
+        self._transport.send(self._create_knxipframe(), addr=self.data_endpoint_addr)
 
-    def create_knxipframe(self) -> KNXIPFrame:
+    def _create_knxipframe(self) -> KNXIPFrame:
         """Create KNX/IP Frame object to be sent to device."""
         return KNXIPFrame.init_from_body(self.device_configuration_request)